+++ src/extension.ts
@@ -3,16 +3,31 @@ import * as fs from 'fs';
 import * as path from 'path';
 import ignore from 'ignore';
 
-type StatusBar = {
+/**
+ * Represents a wrapper around VS Code StatusBarItem,
+ * providing both the created status bar element and
+ * a function to update its text;
+ */
+export type StatusBar = {
     item: vscode.StatusBarItem;
     update: (total: number) => void;
 };
 
-const countTextLines = (text: string): number => {
+/**
+ * Counts the number of non-empty lines in the given text.
+ * @param text - The text as a string.
+ * @returns Count of non-empty lines.
+ */
+export const countTextLines = (text: string): number => {
     return text.split(/\r?\n/).filter((line) => line.trim().length > 0).length;
 };
 
-const countFileLines = (documentPath: string): number => {
+/**
+ * Counts the number of non-empty lines in the given file.
+ * @param documentPath - The path of the file.
+ * @returns Count of non-empty lines in file or 0 in case of error.
+ */
+export const countFileLines = (documentPath: string): number => {
     try {
         const document = fs.readFileSync(documentPath, 'utf8');
         return countTextLines(document);
@@ -21,7 +36,13 @@ const countFileLines = (documentPath: string): number => {
     }
 };
 
-const loadGitignore = (workspacePath: string): ignore.Ignore => {
+/**
+ * Loads `.gitignore` rules from the workspace.
+ * If the `.gitignore` doesn't exist, returns an empty `ignore.Ignore` instance.
+ * @param workspacePath - Absolute path of the workspace directory.
+ * @returns An `ignore.Ignore` instance configured with `.gitignore` rules
+ */
+export const loadGitignore = (workspacePath: string): ignore.Ignore => {
     const ig = ignore();
     const gitignorePath = path.join(workspacePath, '.gitignore');
     if (fs.existsSync(gitignorePath)) {
@@ -31,7 +52,16 @@ const loadGitignore = (workspacePath: string): ignore.Ignore => {
     return ig;
 };
 
-const getTrackedFiles = async (workspacePath: string, ig: ignore.Ignore) => {
+/**
+ * Gets files from directory and filters by gitignore.
+ * @param workspacePath - Absolute path of the workspace directory.
+ * @param ig - A configured `ignore.Ignore` object.
+ * @returns An array of tracked workspace files.
+ */
+export const getTrackedFiles = async (
+    workspacePath: string,
+    ig: ignore.Ignore,
+) => {
     const filePaths = await vscode.workspace.findFiles('**/*');
     return filePaths
         .filter(
@@ -41,7 +71,11 @@ const getTrackedFiles = async (workspacePath: string, ig: ignore.Ignore) => {
         .map((filePath) => filePath.fsPath);
 };
 
-const createStatusBar = (): StatusBar => {
+/**
+ * Creates status bar item and function to set text value of it.
+ * @returns StatusBar object that contains created status bar item and function to update it.
+ */
+export const createStatusBar = (): StatusBar => {
     const statusBarItem = vscode.window.createStatusBarItem(
         vscode.StatusBarAlignment.Left,
         100,
@@ -56,12 +90,27 @@ const createStatusBar = (): StatusBar => {
     };
 };
 
-const workspaceState = {
+/**
+ * Global state manager for the workspace.
+ * Tracks the workspace path, total number of non-empty lines, individual file counts,
+ * gitignore rules. Provides methods to initialize workspace, handle document changes
+ * and handle savings in the `.gitignore` file.
+ */
+export const workspaceState = {
+    /** Current workspace directory absolute path or undefined if no workspace is open */
     path: undefined as string | undefined,
+    /** Total number of non-empty lines in all tracked files in the workspace */
     total: 0,
+    /** A mapping from file paths to the number of non-empty lines in that file */
     fileCounts: {} as Record<string, number>,
+    /** Ignore rules from the `.gitignore` file */
     ig: ignore(),
 
+    /**
+     * Initializes the workspace state.
+     * Loads `.gitignore`, counts lines in all tracked files, updates the status bar.
+     * @param StatusBar - The status bar object used to display total lines.
+     */
     async initialize({
         item: statusBarItem,
         update: updateStatusBar,
@@ -85,6 +134,12 @@ const workspaceState = {
         updateStatusBar(this.total);
     },
 
+    /**
+     * Handles saving a file. If the `.gitignore` file is saved,
+     * reinitializes the workspace state.
+     * @param docPath - The absolute path of the saved document.
+     * @param statusBar - The status bar object to update the display.
+     */
     handleDocumentSave(docPath: string, statusBar: StatusBar) {
         if (!this.path) {
             return;
@@ -95,6 +150,12 @@ const workspaceState = {
         }
     },
 
+    /**
+     * Handles changes to a text document. Updates the file lines count and
+     * total workspace lines count if the document is not ignored.
+     * @param doc - The changed file.
+     * @param statusBar - The status bar object to update the display.
+     */
     handleDocumentChange(doc: vscode.TextDocument, statusBar: StatusBar) {
         if (!this.path || doc.uri.scheme !== 'file') {
             return;
@@ -114,39 +175,54 @@ const workspaceState = {
     },
 };
 
+/**
+ * Counts the number of non-empty lines in the current file and displays
+ * it as a VS Code information message.
+ */
+export const countFileLinesCommand = () => {
+    const editor = vscode.window.activeTextEditor;
+    if (!editor) {
+        vscode.window.showWarningMessage('There is no open file for analysis');
+        return;
+    }
+    const documentPath = editor.document.fileName;
+    vscode.window.showInformationMessage(
+        `Non-empty strings in the file: ${countFileLines(documentPath)}`,
+    );
+};
+
+/**
+ * Counts the number of non-empty lines in the current workspace directory
+ * and displays it as a VS Code information message.
+ */
+export const countWorkspaceLinesCommand = async () => {
+    if (!workspaceState.path) {
+        vscode.window.showWarningMessage('No workspace is open');
+        return;
+    }
+
+    vscode.window.showInformationMessage(
+        `Non-empty strings in the workspace: ${workspaceState.total}`,
+    );
+};
+
+/**
+ * Entry point of the VS Code extension.
+ * Registers commands, initializes workspace tracking,
+ * and subscribes to file change events.
+ */
 export async function activate(context: vscode.ExtensionContext) {
     console.log('Code Line Counter is now active!');
     const statusBar = createStatusBar();
 
     const printFileLines = vscode.commands.registerCommand(
         'code-line-counter.countFileLines',
-        () => {
-            const editor = vscode.window.activeTextEditor;
-            if (!editor) {
-                vscode.window.showWarningMessage(
-                    'There is no open file for analysis',
-                );
-                return;
-            }
-            const documentPath = editor.document.fileName;
-            vscode.window.showInformationMessage(
-                `Non-empty strings in the file: ${countFileLines(documentPath)}`,
-            );
-        },
+        countFileLinesCommand,
     );
 
     const printWorkspaceLines = vscode.commands.registerCommand(
         'code-line-counter.countWorkspaceLines',
-        async () => {
-            if (!workspaceState.path) {
-                vscode.window.showWarningMessage('No workspace is open');
-                return;
-            }
-
-            vscode.window.showInformationMessage(
-                `Non-empty strings in the workspace: ${workspaceState.total}`,
-            );
-        },
+        countWorkspaceLinesCommand,
     );
 
     await workspaceState.initialize(statusBar);
@@ -164,4 +240,11 @@ export async function activate(context: vscode.ExtensionContext) {
     );
 }
 
+/**
+ * Deactivates the extension.
+ *
+ * VS Code automatically disposes of all resources registered through
+ * `context.subscriptions`, including event listeners, status bar items,
+ * and commands. No manual cleanup is needed.
+ */
 export function deactivate() {}